requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
cssbeautifier==1.15.1
Django==4.2.9
django-allauth==65.0.2
django-theme-material-kit==1.0.19
djlint==1.35.2
EditorConfig==0.12.4
ffmpeg-python==0.2.0
future==1.0.0
gunicorn==21.2.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.10
image==1.5.33
jsbeautifier==1.15.1
json5==0.9.25
jwt==1.3.1
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
pycparser==2.22
PyJWT==2.3.0
python-dotenv==1.0.1
PyYAML==6.0.2
qrcode==8.0
regex==2024.9.11
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
str2bool==1.1
tqdm==4.66.5
urllib3==2.2.3
whitenoise==6.6.0
